refactor: replace the constants folder with the shared.ts file in the project root
