--- src/constants/urls.ts
@@ -1,1 +1,0 @@
-export const webinarsUrl = "https://try.estuary.dev/webinar-estuary101-ondemand"
