+++ src/layouts/Product/sections/TakeATour/index.tsx
@@ -1,7 +1,7 @@
 import { StaticImage } from "gatsby-plugin-image"
 import React from "react"
 
-import { webinarsUrl } from "../../../../constants/urls"
+import { webinarsUrl } from "../../../../../shared"
 import { ContainerIcon } from "../../../styles"
 import { ActionLink, Button, Container, ContainerButton, ContainerContent, ContainerIcons, ContainerImage, Description, Title, Wrapper } from "./styles"
